Struct zircon_object::dev::pci::PCIeBusDriver

pub struct PCIeBusDriver { /* fields omitted */ }

PCIE Bus Driver.

Implementations

impl PCIeBusDriver

pub fn add_bus_region(base: u64, size: u64, aspace: PciAddrSpace) -> ZxResult

Add bus region.

pub fn sub_bus_region(base: u64, size: u64, aspace: PciAddrSpace) -> ZxResult

Subtract bus region.

pub fn set_address_translation_provider(
provider: Arc<dyn PCIeAddressProvider>
) -> ZxResult

A PcieAddressProvider translates a BDF address to an address that the system can use to access ECAMs.

pub fn add_root(bus_id: usize, lut: PciIrqSwizzleLut) -> ZxResult

Add a root bus to the driver and attempt to scan it for devices.

pub fn start_bus_driver() -> ZxResult

Start the bus driver.

pub fn get_nth_device(
n: usize
) -> ZxResult<(PcieDeviceInfo, Arc<PcieDeviceKObject>)>

Get the “Nth” device.

impl PCIeBusDriver

pub fn get_config(
&self,
bus_id: usize,
dev_id: usize,
func_id: usize
) -> Option<(Arc<PciConfig>, PhysAddr)>

Get a device’s config.

pub fn link_device_to_upstream(
&self,
down: Arc<dyn IPciNode>,
up: Weak<dyn IPciNode>
)

Link a device to an upstream node.

pub fn find_legacy_irq_handler(
&self,
irq_id: usize
) -> ZxResult<Arc<SharedLegacyIrqHandler>>

Find the legacy IRQ handler.