Struct zircon_object::object::Rights

pub struct Rights { /* private fields */ }

Rights are associated with handles and convey privileges to perform actions on either the associated handle or the object associated with the handle.

Implementations

impl Rights

pub const DUPLICATE: Self = _

Allows handle duplication via zx_handle_duplicate().

pub const TRANSFER: Self = _

Allows handle transfer via zx_channel_write().

pub const READ: Self = _

Allows reading of data from containers (channels, sockets, VM objects, etc). Allows mapping as readable if Rights::MAP is also present.

pub const WRITE: Self = _

Allows writing of data to containers (channels, sockets, VM objects, etc). Allows mapping as writeable if Rights::MAP is also present.

pub const EXECUTE: Self = _

Allows mapping as executable if Rights::MAP is also present.

pub const MAP: Self = _

Allows mapping of a VM object into an address space.

pub const GET_PROPERTY: Self = _

Allows property inspection via zx_object_get_property().

pub const SET_PROPERTY: Self = _

Allows property modification via zx_object_set_property().

pub const ENUMERATE: Self = _

Allows enumerating child objects via zx_object_get_info() and zx_object_get_child().

pub const DESTROY: Self = _

Allows termination of task objects via zx_task_kill().

pub const SET_POLICY: Self = _

Allows policy modification via zx_job_set_policy().

pub const GET_POLICY: Self = _

Allows policy inspection via zx_job_get_policy().

pub const SIGNAL: Self = _

Allows use of zx_object_signal().

pub const SIGNAL_PEER: Self = _

Allows use of zx_object_signal_peer().

pub const WAIT: Self = _

Allows use of zx_object_wait_one(), zx_object_wait_many(), and other waiting primitives.

pub const INSPECT: Self = _

Allows inspection via zx_object_get_info().

pub const MANAGE_JOB: Self = _

Allows creation of processes, subjobs, etc.

pub const MANAGE_PROCESS: Self = _

Allows creation of threads, etc.

pub const MANAGE_THREAD: Self = _

Allows suspending/resuming threads, etc.

pub const APPLY_PROFILE: Self = _

Not used.

pub const SAME_RIGHTS: Self = _

Used to duplicate a handle with the same rights.

pub const BASIC: Self = _

TRANSFER | DUPLICATE | WAIT | INSPECT

pub const IO: Self = _

READ ｜ WRITE

pub const PROPERTY: Self = _

GET_PROPERTY ｜ SET_PROPERTY

pub const POLICY: Self = _

GET_POLICY ｜ SET_POLICY

pub const DEFAULT_CHANNEL: Self = _

BASIC & !Self::DUPLICATE | IO | SIGNAL | SIGNAL_PEER

pub const DEFAULT_PROCESS: Self = _

BASIC | IO | PROPERTY | ENUMERATE | DESTROY | SIGNAL | MANAGE_PROCESS | MANAGE_THREAD

pub const DEFAULT_THREAD: Self = _

BASIC | IO | PROPERTY | DESTROY | SIGNAL | MANAGE_THREAD

pub const DEFAULT_VMO: Self = _

BASIC | IO | PROPERTY | MAP | SIGNAL

pub const DEFAULT_VMAR: Self = _

BASIC | WAIT

pub const DEFAULT_JOB: Self = _

BASIC | IO | PROPERTY | POLICY | ENUMERATE | DESTROY | SIGNAL | MANAGE_JOB | MANAGE_PROCESS | MANAGE_THREAD

pub const DEFAULT_RESOURCE: Self = _

TRANSFER | DUPLICATE | WRITE | INSPECT

pub const DEFAULT_DEBUGLOG: Self = _

BASIC | WRITE | SIGNAL

pub const DEFAULT_SUSPEND_TOKEN: Self = _

TRANSFER | INSPECT

pub const DEFAULT_PORT: Self = _

(BASIC & !WAIT) | IO

pub const DEFAULT_TIMER: Self = _

BASIC | WRITE | SIGNAL

pub const DEFAULT_EVENT: Self = _

BASIC | SIGNAL

pub const DEFAULT_EVENTPAIR: Self = _

BASIC | SIGNAL ｜ SIGNAL_PEER

pub const DEFAULT_FIFO: Self = _

BASIC | IO | SIGNAL | SIGNAL_PEER

pub const DEFAULT_SOCKET: Self = _

BASIC | IO | PROPERTY | SIGNAL | SIGNAL_PEER

pub const DEFAULT_STREAM: Self = _

BASIC | PROPERTY | SIGNAL

pub const DEFAULT_BTI: Self = _

(BASIC & !WAIT) | IO | MAP

pub const DEFAULT_INTERRUPT: Self = _

BASIC | IO | SIGNAL

pub const DEFAULT_DEVICE: Self = _

BASIC | IO

pub const DEFAULT_PCI_INTERRUPT: Self = _

BASIC | IO | SIGNAL

pub const DEFAULT_EXCEPTION: Self = _

TRANSFER | PROPERTY | INSPECT

pub const DEFAULT_GUEST: Self = _

TRANSFER | DUPLICATE | WRITE | INSPECT | MANAGE_PROCESS

pub const DEFAULT_VCPU: Self = _

BASIC | IO | EXECUTE | SIGNAL

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u32

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u32) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.