Struct zircon_object::task::CurrentThread

pub struct CurrentThread(_);

A handle to current thread.

This is a wrapper of Thread that provides additional methods for the thread runner. It can only be obtained from the argument of thread_fn in a new thread started by Thread::start.

It will terminate current thread on drop.

Implementations

impl CurrentThread

pub fn exit(&self)

Exit the current thread.

The thread do not terminate immediately when exited. It is just made dying. It will terminate after some cleanups on this struct drop.

pub fn wait_for_run(&self) -> impl Future<Output = Box<UserContext>>

Wait until the thread is ready to run (not suspended), and then take away its context to run the thread.

pub fn end_running(&self, context: Box<UserContext>)

The thread ends running and takes back the context.

pub fn with_context<T, F>(&self, f: F) -> T where
F: FnOnce(&mut UserContext) -> T,

Access saved context of current thread.

Will panic if the context is not availiable.

pub async fn blocking_run<F, T, FT>(
&self,
future: F,
state: ThreadState,
deadline: Duration,
cancel_token: Option<Receiver<()>>
) -> ZxResult<T> where
F: Future<Output = FT> + Unpin,
FT: IntoResult<T>,

Run async future and change state while blocking.

pub async fn handle_exception(&self, type_: ExceptionType)

Create an exception on this thread and wait for the handling.