Trait zircon_object::task::IntoResult

pub trait IntoResult<T> {
    fn into_result(self) -> ZxResult<T>;
}

into_result returns Self if the type parameter is already a ZxResult, otherwise wraps the value in an Ok.

Used to implement Thread::blocking_run, which takes a future whose Output may or may not be a ZxResult.

Required methods

fn into_result(self) -> ZxResult<T>

Performs the conversion.

Loading content...