Struct zircon_object::vm::VmAddressRegion

pub struct VmAddressRegion { /* fields omitted */ }

Virtual Memory Address Regions

Implementations

impl VmAddressRegion

pub fn new_root() -> Arc<Self>

Create a new root VMAR.

pub fn new_kernel() -> Arc<Self>

Create a kernel root VMAR.

pub fn new_guest() -> Arc<Self>

Create a VMAR for guest physical memory.

pub fn allocate_at(
self: &Arc<Self>,
offset: usize,
len: usize,
flags: VmarFlags,
align: usize
) -> ZxResult<Arc<Self>>

Create a child VMAR at the offset.

pub fn allocate(
self: &Arc<Self>,
offset: Option<usize>,
len: usize,
flags: VmarFlags,
align: usize
) -> ZxResult<Arc<Self>>

Create a child VMAR with optional offset.

pub fn map_at(
&self,
vmar_offset: usize,
vmo: Arc<VmObject>,
vmo_offset: usize,
len: usize,
flags: MMUFlags
) -> ZxResult<VirtAddr>

Map the vmo into this VMAR at given offset.

pub fn map(
&self,
vmar_offset: Option<usize>,
vmo: Arc<VmObject>,
vmo_offset: usize,
len: usize,
flags: MMUFlags
) -> ZxResult<VirtAddr>

Map the vmo into this VMAR.

pub fn map_ext(
&self,
vmar_offset: Option<usize>,
vmo: Arc<VmObject>,
vmo_offset: usize,
len: usize,
permissions: MMUFlags,
flags: MMUFlags,
overwrite: bool,
map_range: bool
) -> ZxResult<VirtAddr>

Map the vmo into this VMAR.

pub fn unmap(&self, addr: VirtAddr, len: usize) -> ZxResult

Unmaps all VMO mappings and destroys all sub-regions within the absolute range including addr and ending before exclusively at addr + len. Any sub-region that is in the range must be fully in the range (i.e. partial overlaps are an error). If a mapping is only partially in the range, the mapping is split and the requested portion is unmapped.

pub fn protect(&self, addr: usize, len: usize, flags: MMUFlags) -> ZxResult

Change protections on a subset of the region of memory in the containing address space. If the requested range overlaps with a subregion, protect() will fail.

pub fn destroy(self: &Arc<Self>) -> ZxResult

Unmap all mappings within the VMAR, and destroy all sub-regions of the region.

pub fn clear(&self) -> ZxResult

Unmap all mappings and destroy all sub-regions of VMAR.

pub fn table_phys(&self) -> PhysAddr

Get physical address of the underlying page table.

pub fn addr(&self) -> usize

Get start address of this VMAR.

pub fn is_dead(&self) -> bool

Whether this VMAR is dead.

pub fn is_alive(&self) -> bool

Whether this VMAR is alive.

pub fn get_info(&self) -> VmarInfo

Get information of this VmAddressRegion

pub fn get_flags(&self) -> VmarFlags

Get VmarFlags of this VMAR.

pub fn dump(&self)

Dump all mappings recursively.

pub fn vdso_base_addr(&self) -> Option<usize>

Get base address of vdso.

pub fn handle_page_fault(&self, vaddr: VirtAddr, flags: MMUFlags) -> ZxResult

Handle page fault happened on this VMAR.

The fault virtual address is vaddr and the reason is in flags.

pub fn fork_from(&self, src: &Arc<Self>) -> ZxResult

Clone the entire address space and VMOs from source VMAR. (For Linux fork)

pub fn get_task_stats(&self) -> TaskStatsInfo

Returns statistics about memory used by a task.

pub fn read_memory(&self, vaddr: usize, buf: &mut [u8]) -> ZxResult<usize>

Read from address space.

Return the actual number of bytes read.

pub fn write_memory(&self, vaddr: usize, buf: &[u8]) -> ZxResult<usize>

Write to address space.

Return the actual number of bytes written.

pub fn find_mapping(&self, vaddr: usize) -> Option<Arc<VmMapping>>

Find mapping of vaddr