Module zircon_object::signal

Objects for signaling and waiting.

Structs

Event

Signalable event for concurrent programming

EventPair

Mutually signalable pair of events for concurrent programming

Futex

A primitive for creating userspace synchronization tools.

PacketGuestBell

PacketGuestIo

PacketGuestMem

PacketGuestVcpu

PacketGuestVcpuInterrupt

PacketGuestVcpuStartup

PacketInterrupt

PacketSignal

Port

Signaling and mailbox primitive

PortOptions

If you need this port to be bound to an interrupt, pass BIND_TO_INTERRUPT to options, otherwise it should be 0.

PortPacket

A packet sent through a port.

PortPacketRepr

A high-level representation of a packet sent through a port.

Timer

An object that may be signaled at some point in the future

Enums

PacketGuestVcpuType

PacketType

The type of a packet.

PayloadRepr

A high-level representation of a packet payload.

Slack

Slack specifies how much a timer or event is allowed to deviate from its deadline.

Type Definitions

PacketUser

Unions

PacketGuestVcpuData

Payload

The data carried by a packet