Struct zircon_object::task::Job

pub struct Job { /* fields omitted */ }

Control a group of processes

A job is a group of processes and possibly other (child) jobs. Jobs are used to track privileges to perform kernel operations (i.e., make various syscalls, with various options), and track and limit basic resource (e.g., memory, CPU) consumption. Every process belongs to a single job. Jobs can also be nested, and every job except the root job also belongs to a single (parent) job.

DESCRIPTION

A job is an object consisting of the following:

a reference to a parent job

a set of child jobs (each of whom has this job as parent)

a set of member processes

a set of policies

Jobs control “applications” that are composed of more than one process to be controlled as a single entity.

Implementations

impl Job

pub fn root() -> Arc<Self>

Create the root job.

pub fn create_child(self: &Arc<Self>) -> ZxResult<Arc<Self>>

Create a new child job object.

pub fn policy(&self) -> JobPolicy

Get the policy of the job.

pub fn parent(&self) -> Option<Arc<Self>>

Get the parent job.

pub fn set_policy_basic(
&self,
options: SetPolicyOptions,
policies: &[BasicPolicy]
) -> ZxResult

Sets one or more security and/or resource policies to an empty job.

The job’s effective policies is the combination of the parent’s effective policies and the policies specified in policy.

After this call succeeds any new child process or child job will have the new effective policy applied to it.

pub fn set_policy_timer_slack(&self, policy: TimerSlackPolicy) -> ZxResult

Sets timer slack policy to an empty job.

pub fn get_info(&self) -> JobInfo

Get information of this job.

pub fn check_root_job(&self) -> ZxResult

Check whether this job is root job.

pub fn process_ids(&self) -> Vec<KoID>

Get KoIDs of Processes.

pub fn children_ids(&self) -> Vec<KoID>

Get KoIDs of children Jobs.

pub fn is_empty(&self) -> bool

Return true if this job has no processes and no child jobs.