Trait zircon_object::task::Task

pub trait Task: Sync + Send {
    fn kill(&self);
    fn suspend(&self);
    fn resume(&self);
    fn exceptionate(&self) -> Arc<Exceptionate>;
    fn debug_exceptionate(&self) -> Arc<Exceptionate>;
}

Task (Thread, Process, or Job)

Required Methods

fn kill(&self)

Kill the task. The task do not terminate immediately when killed. It will terminate after all its children are terminated or some cleanups are finished.

fn suspend(&self)

Suspend the task. Currently only thread or process handles may be suspended.

fn resume(&self)

Resume the task

fn exceptionate(&self) -> Arc<Exceptionate>

Get the exceptionate.

fn debug_exceptionate(&self) -> Arc<Exceptionate>

Get the debug exceptionate.