Trait zircon_object::util::elf_loader::VmarExt

pub trait VmarExt {
    fn load_from_elf(&self, elf: &ElfFile<'_>) -> ZxResult<Arc<VmObject>>;

    fn map_from_elf(&self, elf: &ElfFile<'_>, vmo: Arc<VmObject>) -> ZxResult;
}

Extensional ELF loading methods for VmAddressRegion.

Required methods

fn load_from_elf(&self, elf: &ElfFile<'_>) -> ZxResult<Arc<VmObject>>

Create VMObject from all LOAD segments of elf and map them to this VMAR. Return the first VMObject.

fn map_from_elf(&self, elf: &ElfFile<'_>, vmo: Arc<VmObject>) -> ZxResult

Same as load_from_elf, but the vmo is an existing one instead of a lot of new ones.