Struct zircon_object::vm::VmObject

pub struct VmObject { /* private fields */ }

Virtual memory containers

A Virtual Memory Object (VMO) represents a contiguous region of virtual memory that may be mapped into multiple address spaces.

Implementations

impl VmObject

pub fn new_paged(pages: usize) -> Arc<Self>

Create a new VMO backing on physical memory allocated in pages.

pub fn new_paged_with_resizable(resizable: bool, pages: usize) -> Arc<Self>

Create a new VMO, which can be resizable, backing on physical memory allocated in pages.

pub fn new_physical(paddr: PhysAddr, pages: usize) -> Arc<Self>

Create a new VMO representing a piece of contiguous physical memory.

pub fn new_contiguous(pages: usize, align_log2: usize) -> ZxResult<Arc<Self>>

Create a VM object referring to a specific contiguous range of physical frame.

pub fn create_child(
self: &Arc<Self>,
resizable: bool,
offset: usize,
len: usize
) -> ZxResult<Arc<Self>>

Create a child VMO.

pub fn create_slice(
self: &Arc<Self>,
offset: usize,
p_size: usize
) -> ZxResult<Arc<Self>>

Create a child slice as an VMO

pub fn set_len(&self, len: usize) -> ZxResult

Set the length of this VMO if resizable.

pub fn set_content_size_and_resize(
&self,
size: usize,
zero_until_offset: usize
) -> ZxResult<usize>

Set the size of the content stored in the VMO in bytes, resize vmo if needed

pub fn content_size(&self) -> usize

Get the size of the content stored in the VMO in bytes.

pub fn set_content_size(&self, size: usize) -> ZxResult

Get the size of the content stored in the VMO in bytes.

pub fn get_info(&self) -> VmoInfo

Get information of this VMO.

pub fn set_cache_policy(&self, policy: CachePolicy) -> ZxResult

Set the cache policy.

pub fn append_mapping(&self, mapping: Weak<VmMapping>)

Append a mapping to the VMO’s mapping list.

pub fn remove_mapping(&self, mapping: Weak<VmMapping>)

Remove a mapping from the VMO’s mapping list.

pub fn share_count(&self) -> usize

Returns an estimate of the number of unique VmAspaces that this object is mapped into.

pub fn is_resizable(&self) -> bool

Returns true if the object size can be changed.

pub fn is_contiguous(&self) -> bool

Returns true if the object is backed by a contiguous range of physical memory.